Toggling is fine, but for testing, you need absolutes.
This adds A.reset_informant for testing (only). It is used in an
ensure on test_reports_get_made_from_setter_methods to clean up after
itself. You could also use it in teardown, but meh.

This also only turns on the informant in the one test that actually
uses it (i.e., the one test that failed when I removed the toggle from
setup).